migration: increase listening socket backlog
When using live migration with multifd and a large number of channels,
the incoming tcp connections have to be held in the backlog.
The number of these is set to the number of multifd channels
in case of deferred migration ('-incoming defer'). But the backlog
is set to 1 when the number of multifd channels to use when Qemu
starts is unknown. This happens when Qemu is started with '-incoming
tcp:*:4444' for example and the number of multifd channels is being set
later by hmp/qmp command.

OraBug: 32792545

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Karl Heubaum <karl.heubaum@oracle.com>
Reviewed-by: David Edmondson <david.edmondson@oracle.com>
(cherry picked from commit 77c34e39feb223dfabe92c0c4264ef80aadb2cff)

Jira: https://jira.oci.oraclecorp.com/browse/LINUX-13266

Signed-off-by: Karl Heubaum <karl.heubaum@oracle.com>