multifd: Protect multifd_send_sync_main() calls
We only need to do that on the ram_save_iterate() call on sending and
on destination when we get a RAM_SAVE_FLAG_EOS.

In setup() and complete() we need to synch in both new and old cases,
so don't add a check there.

Signed-off-by: Juan Quintela <quintela@redhat.com>
Reviewed-by: Dr. David Alan Gilbert <dgilbert@redhat.com>
Acked-by: Peter Xu <peterx@redhat.com>

Remove the wrappers that we take out on patch 5.

 Conflicts:
       migration/ram.c - we use rs->f to reference a channel
       as we do not have commit commit 7f401b8044
       ("migration: Drop rs->f").

(cherry picked from commit b05292c237030343516d073b1a1e5f49ffc017a8)

Orabug: 35636284

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>