oslib-posix: initialize backend memory objects in parallel
QEMU initializes preallocated backend memory as the objects are parsed from
the command line. This is not optimal in some cases (e.g. multiple sockets)
because the memory objects are initialized in series.

Allow the initialization to occur in parallel. The performance increase is
significant and scales with the number of sockets. On a 2 socket Skylake VM
with 128GB and 2 init threads per socket (256GB total), the memory init time
decreases from ~27 seconds to ~14 seconds.

Orabug: 32555402

Signed-off-by: Mark Kanda <mark.kanda@oracle.com>
Tested-by: Wim Ten Have <wim.ten.have@oracle.com>
Tested-by: Rose Wang <rose.wang@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>
(cherry picked from commit 9470924f47f51c24f5bc7861a01b4d54d2094997)

Signed-off-by: Mark Kanda <mark.kanda@oracle.com>
Reviewed-by: Karl Heubaum <karl.heubaum@oracle.com>

Conflicts:
    softmmu/vl.c
    util/oslib-posix.c

- Addressed misc contextual conflicts
- Added return handling for wait_mem_prealloc()