pcie: Do not update hotplugged device power in RUN_STATE_INMIGRATE state
If Windows source VM boots from the system disk hot plugged in
RUN_STATE_PRELAUNCH state, the system disk needs to be
hotplugged into the Windows receiving VM before live migration.
The running state of the receiving VM is RUN_STATE_INMIGRATE. In
this case, the slot control power bit is set as OFF, and the power
of the system disk is updated to OFF.

As a result, Windows receiving VM hangs after the migration is
done since the power of the system disk is OFF. This patch skips
updating the power of the device that is hot-plugged in
RUN_STATE_INMIGRATE state.

Linux guest sets the slot power bit of the device to ON after the
migration is done, so Linux guest doesn't run into this issue.

Fixes: f08fb5b8c934 ("pcie: Do not set power state for some hot-plugged devices")

Orabug: 35055290

Signed-off-by: Annie Li <annie.li@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>
(cherry picked from commit 98f1521076490c186705243cac7e2e8e194f8f6a)

Orabug: 35576284

Signed-off-by: Annie Li <annie.li@oracle.com>
Reviewed-by: Mark Kanda <mark.kanda@oracle.com>