migration: Add switchover ack capability
Migration downtime estimation is calculated based on bandwidth and
remaining migration data. This assumes that loading of migration data in
the destination takes a negligible amount of time and that downtime
depends only on network speed.

While this may be true for RAM, it's not necessarily true for other
migrated devices. For example, loading the data of a VFIO device in the
destination might require from the device to allocate resources, prepare
internal data structures and so on. These operations can take a
significant amount of time which can increase migration downtime.

This patch adds a new capability "switchover ack" that prevents the
source from stopping the VM and completing the migration until an ACK
is received from the destination that it's OK to do so.

This can be used by migrated devices in various ways to reduce downtime.
For example, a device can send initial precopy metadata to pre-allocate
resources in the destination and use this capability to make sure that
the pre-allocation is completed before the source VM is stopped, so it
will have full effect.

This new capability relies on the return path capability to communicate
from the destination back to the source.

The actual implementation of the capability will be added in the
following patches.

Signed-off-by: Avihai Horon <avihaih@nvidia.com>
Reviewed-by: Peter Xu <peterx@redhat.com>
Acked-by: Markus Armbruster <armbru@redhat.com>
Tested-by: YangHang Liu <yanghliu@redhat.com>
Acked-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Cédric Le Goater <clg@redhat.com>
(cherry picked from commit 6574232fff6a810e1882c98aa7dcf23d6afe9a31)

Jira: https://jira.oci.oraclecorp.com/browse/LINUX-13956
Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Reviewed-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>

Conflicts:
	migration/options.c
	migration/options.h

There was a big rework in Qemu 8.0 i.e. commit 1f0776f1c033 ("migration:
Create options.c") that reorganized migration parameters and that placed
most of its parameterization code in options.c. The same exact chunks are
applied just the same in migration.{c,h}.

A few ammends are made for to
be compatible with 7.2: 1) intend to the start of the property otherwise we
fail compilation with:

qemu/scripts/qapi-gen.py: In file included from ../qapi/qapi-schema.json:79:
../qapi/migration.json:482:1: unexpected de-indent (expected at least 17 spaces)

2) a few renames due to the lack of that big cleanup of migration
options: new_caps => cap_list and capabilities => enabled_capablities