vfio/common: Add an option to relax vIOMMU usage
Add an option 'x-orcl-migration-iommu-pt' to VFIO that allows it to relax
whether the vIOMMU migration is allowed, in case there was no gIOMMU
mappings done. In those cases, we use the boot direct map to establish
dirty tracking, as opposed to stressing the limits of device.

This is meant to be used for managed VMs cases where we know what is being
used by the VM, and thus we can relax the usage boundaries. Default is
'disabled', because the default on Qemu is to *not* trust the guests.

JIRA: https://jira.oci.oraclecorp.com/browse/LINUX-13969
Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Reviewed-by: Karl Heubaum <karl.heubaum@oracle.com>