vfio/migration: Fail adding device with enable-migration=on and existing blocker
If a device with enable-migration=on is added and it causes a migration
blocker, adding the device should fail with a proper error.

This is not the case with multiple device migration blocker when the
blocker already exists. If the blocker already exists and a device with
enable-migration=on is added which causes a migration blocker, adding
the device will succeed.

Fix it by failing adding the device in such case.

Fixes: 8bbcb64a71d8 ("vfio/migration: Make VFIO migration non-experimental")
Signed-off-by: Avihai Horon <avihaih@nvidia.com>
Reviewed-by: Cédric Le Goater <clg@redhat.com>
Signed-off-by: Cédric Le Goater <clg@redhat.com>
(cherry picked from commit 8118349b1bc5c12e4017ddd7eb563fce752a5a9c)

JIRA: https://jira.oci.oraclecorp.com/browse/LINUX-14257

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Reviewed-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>