migration: Allow user to specify available switchover bandwidth
Migration bandwidth is a very important value to live migration.  It's
because it's one of the major factors that we'll make decision on when to
switchover to destination in a precopy process.

This value is currently estimated by QEMU during the whole live migration
process by monitoring how fast we were sending the data.  This can be the
most accurate bandwidth if in the ideal world, where we're always feeding
unlimited data to the migration channel, and then it'll be limited to the
bandwidth that is available.

However in reality it may be very different, e.g., over a 10Gbps network we
can see query-migrate showing migration bandwidth of only a few tens of
MB/s just because there are plenty of other things the migration thread
might be doing.  For example, the migration thread can be busy scanning
zero pages, or it can be fetching dirty bitmap from other external dirty
sources (like vhost or KVM).  It means we may not be pushing data as much
as possible to migration channel, so the bandwidth estimated from "how many
data we sent in the channel" can be dramatically inaccurate sometimes.

With that, the decision to switchover will be affected, by assuming that we
may not be able to switchover at all with such a low bandwidth, but in
reality we can.

The migration may not even converge at all with the downtime specified,
with that wrong estimation of bandwidth, keeping iterations forever with a
low estimation of bandwidth.

The issue is QEMU itself may not be able to avoid those uncertainties on
measuing the real "available migration bandwidth".  At least not something
I can think of so far.

One way to fix this is when the user is fully aware of the available
bandwidth, then we can allow the user to help providing an accurate value.

For example, if the user has a dedicated channel of 10Gbps for migration
for this specific VM, the user can specify this bandwidth so QEMU can
always do the calculation based on this fact, trusting the user as long as
specified.  It may not be the exact bandwidth when switching over (in which
case qemu will push migration data as fast as possible), but much better
than QEMU trying to wildly guess, especially when very wrong.

A new parameter "avail-switchover-bandwidth" is introduced just for this.
So when the user specified this parameter, instead of trusting the
estimated value from QEMU itself (based on the QEMUFile send speed), it
trusts the user more by using this value to decide when to switchover,
assuming that we'll have such bandwidth available then.

Note that specifying this value will not throttle the bandwidth for
switchover yet, so QEMU will always use the full bandwidth possible for
sending switchover data, assuming that should always be the most important
way to use the network at that time.

This can resolve issues like "unconvergence migration" which is caused by
hilarious low "migration bandwidth" detected for whatever reason.

Reported-by: Zhiyi Guo <zhguo@redhat.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>
Reviewed-by: Juan Quintela <quintela@redhat.com>
Signed-off-by: Peter Xu <peterx@redhat.com>
Signed-off-by: Juan Quintela <quintela@redhat.com>
Message-ID: <20231010221922.40638-1-peterx@redhat.com>
(cherry picked from commit 8b2395970aa3beab91b98dda89c7ed471e65ad25)

 Conflicts:
	migration/migration-hmp-cmds.c
        - we use ./monitor/hmp-cmds.c for changes since we dont
        have commit 119f50ce30f1dfdfd33e4ec7455b147834c794d5
        "migration: Move HMP commands from monitor/ to migration/"

	migration/options.c
	migration/options.h
        - these files also do not exist in this version since
          we don't have commit 1f0776f1c033 (""migration: Create
          options.c""), apply to migration.c and migration.h.

	migration/migration.c
        - semantic of migration_transferred_bytes() is different
        since we do not have commit f57e5a6ce50f10941399014c628db8713c2bc3f9
        "    migration: migration_transferred_bytes() don't need the
        QEMUFile".
        - we do not have 0743f41fd293b033de289c52faa45d06a5b1d544
        "    migration: migration_rate_limit_reset() don't need the
        QEMUFile", use existing semantics.
        - use ram_counters instead of ram_stats;
        - we do not have f5da8ba4777f3ab88bfee82bbf8261d6295a26a2
        "    migration: Create migrate_downtime_limit() function".

	migration/trace-events
        - no commit d8b71d96b336054ac3cd0c9351b0cb75c47281ad
        "    migration: spelling fixes", replacing with upstream
        version.

	qapi/migration.json
        - mostly context sice we do not have commit
        a937b6aa739f65f2cae2ad9a7eb65a309ad2a359 "    qapi: Reformat doc
        comments to conform to current conventions" and need to follow
        old spacing requirment for qapi schemas.
        - fix in place "Since <version>" from upstream 8.2 to 7.2.

Orabug: 35636284

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>