migration/multifd: Clarify Error usage in multifd_channel_connect
The function is currently called from two sites, one always gives it a
NULL Error and the other always gives it a non-NULL Error.

In the non-NULL case, all it does it trace the error and return. One
of the callers already have tracing, add a tracepoint to the other and
stop passing the error into the function.

Cc: Markus Armbruster <armbru@redhat.com>
Signed-off-by: Fabiano Rosas <farosas@suse.de>
Reviewed-by: Philippe Mathieu-Daudé <philmd@linaro.org>
Reviewed-by: Juan Quintela <quintela@redhat.com>
Signed-off-by: Juan Quintela <quintela@redhat.com>
Message-ID: <20231012134343.23757-4-farosas@suse.de>
(cherry picked from commit 967e3889874b1116090a60c0cb43157130bdbd16)

Conflicts:
    migration/multifd.c we do not have commit 0e92f6444834a84aefab8f31b4f6a04c0dfed801
    ("migration/multifd: Remove direct "socket" references"), so we use
    sioc channel reference instead of ioc.

    We do not have commit bca762c2b90979e8cf1f725cec3ddf888b7b898b
    ("multifd: Remove some redundant code"), but this cherry-pick
    incorporates a part of it in multifd_new_send_channel_async()
    by removing "goto cleanup" after multifd_channel_connect().

Orabug: 35636284

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>