migration: Use the number of transferred bytes directly
We only use migration_transferred_bytes() to calculate the rate_limit,
for that we don't need to flush whatever is on the qemu_file buffer.
Remember that the buffer is really small (normal case is 32K if we use
iov's can be 64 * TARGET_PAGE_SIZE), so this is not relevant to
calculations.

Reviewed-by: Fabiano Rosas <farosas@suse.de>
Signed-off-by: Juan Quintela <quintela@redhat.com>
Message-ID: <20231025091117.6342-5-quintela@redhat.com>
(cherry picked from commit 737840e2c6ea76896ffd8d3f5474fd4a4aee62d6)

 Conflicts:
	migration/migration-stats.c
        - we use ram_counters instead of mig_stats.

        - we dont have commit 67c31c9c1af1bb8f7df8275cc8731629e2690f89
        ("migration: Don't abuse qemu_file transferred for RDMA").
        Thus we dont account rdma bytes in the total of bytes
        transferred during live migration. If RDMA will be used, fix
        this.

Orabug: 35636284

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>