vfio/common: Probe type1 iommu dirty tracking support
The type1 VFIO migration cap advertises support for the interface.
But under the hood it is marking every with write permission as dirty.

Probe a UEK only capability that reuses the same interface but whether
it uses IOMMU pagetables to tell if it's dirty or not. In doing so,
it also relaxes the requirement for device dirty page tracking (be it
no supported by device or disabled by command line) in order to leverage
the type1 extension.

The linux-headers aren't updated as this won't be upstream and not meant
to last as we will be using IOMMUFD in future major releases.

Orabug: 36024839

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Tested-by: Gerald Gibson <gerald.gibson@oracle.com>
Tested-by: Juan Perez <juan.z.perez@oracle.com>
Reviewed-by: Karl Heubaum <karl.heubaum@oracle.com>