migration/multifd: Postpone reset of MultiFDPages_t
Now we reset MultiFDPages_t object in the multifd sender thread in the
middle of the sending job.  That's not necessary, because the "*pages"
struct will not be reused anyway until pending_job is cleared.

Move that to the end after the job is completed, provide a helper to reset
a "*pages" object.  Use that same helper when free the object too.

This prepares us to keep using p->pages in the follow up patches, where we
may drop p->normal[].

Reviewed-by: Fabiano Rosas <farosas@suse.de>
Link: https://lore.kernel.org/r/20240202102857.110210-5-peterx@redhat.com
Signed-off-by: Peter Xu <peterx@redhat.com>
(cherry picked from commit 836eca47f62f9f6d5b8e9b6fedfc3539775c4e2e)

 Conflicts:
	migration/multifd.c
        - context sine we do not have commit dca1bc7f24d2fa227f0b787f85f3cc67006e67bf
          "migration/multifd: Remove MultiFDPages_t::packet_num".

Orabug: 34131170

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>