migration/multifd: Move header prepare/fill into send_prepare()
This patch redefines the interfacing of ->send_prepare().  It further
simplifies multifd_send_thread() especially on zero copy.

Now with the new interface, we require the hook to do all the work for
preparing the IOVs to send.  After it's completed, the IOVs should be ready
to be dumped into the specific multifd QIOChannel later.

So now the API looks like:

  p->pages ----------->  send_prepare() -------------> IOVs

This also prepares for the case where the input can be extended to even not
any p->pages.  But that's for later.

This patch will achieve similar goal of what Fabiano used to propose here:

https://lore.kernel.org/r/20240126221943.26628-1-farosas@suse.de

However the send() interface may not be necessary.  I'm boldly attaching a
"Co-developed-by" for Fabiano.

Co-developed-by: Fabiano Rosas <farosas@suse.de>
Reviewed-by: Fabiano Rosas <farosas@suse.de>
Link: https://lore.kernel.org/r/20240202102857.110210-14-peterx@redhat.com
Signed-off-by: Peter Xu <peterx@redhat.com>
(cherry picked from commit 25a1f8787597f6906b151b2f73ae6cc92a31de57)

 Conflicts:
	migration/multifd.c
        - no commit b4bc342c766640e0cb8a0b72f71e0ee5545fb790
          "migration: Move migrate_use_zero_copy_send() to options.c"
          hence use of migrate_use_zero_copy_send preserved.
        - context in multifd_send_thread() since no commit 1b1f4ab69c41279a45ccd0d3178e83471e6e4ec1
          "migration: save/delete migration thread info".

Orabug: 34131170

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>