migration/multifd: solve zero page causing multiple page faults
Implemented recvbitmap tracking of received pages in multifd.

If the zero page appears for the first time in the recvbitmap, this
page is not checked and set.

If the zero page has already appeared in the recvbitmap, there is no
need to check the data but directly set the data to 0, because it is
unlikely that the zero page will be migrated multiple times.

Signed-off-by: Yuan Liu <yuan1.liu@intel.com>
Reviewed-by: Fabiano Rosas <farosas@suse.de>
Link: https://lore.kernel.org/r/20240401154110.2028453-2-yuan1.liu@intel.com
[peterx: touch up the comment, as the bitmap is used outside postcopy now]
Signed-off-by: Peter Xu <peterx@redhat.com>
(cherry picked from commit 5ef7e26bdb7eda10d6d5e1b77121be9945e5e550)

 Conflicts:
	include/exec/ramblock.h
        We do not have commit c2d5c4a7cb7f59125c34e5b883d5d91336761661
        "migration/ram: Add outgoing 'mapped-ram' migration"
        and corresponding fields (file_bmap, bitmap_offset,
        pages_offset) are not present in RAMBlock struct.

        not a conflict: added ram.h into include list of multifd-zsdt.c
        as is does not compile without it.

Orabug: 36727051

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>