migration/multifd: Make MultiFDPages_t:offset a flexible array member
We're about to use MultiFDPages_t from inside the MultiFDSendData
payload union, which means we cannot have pointers to allocated data
inside the pages structure, otherwise we'd lose the reference to that
memory once another payload type touches the union. Move the offset
array into the end of the structure and turn it into a flexible array
member, so it is allocated along with the rest of MultiFDSendData in
the next patches.

Note that other pointers, such as the ramblock pointer are still fine
as long as the storage for them is not owned by the migration code and
can be correctly released at some point.

Reviewed-by: Peter Xu <peterx@redhat.com>
Signed-off-by: Fabiano Rosas <farosas@suse.de>
(cherry picked from commit 0e427da096d46860c2a5755295d1e05647e90154)

JIRA: https://jira.oci.oraclecorp.com/browse/LINUX-15752

Conflicts:
	migration/multifd.c: Contextual due to missing commits
	f427d90b9898 ("migration/multifd: Support outgoing mapped-ram
	stream format") and 11dd7be57524 ("migration/multifd: Remove
	p->quit from recv side").

Reviewed-by: Liam Merwick <liam.merwick@oracle.com>
Signed-off-by: Maciej S. Szmigiero <maciej.szmigiero@oracle.com>