migration/multifd: Replace multifd_send_state->pages with client data
Multifd currently has a simple scheduling mechanism that distributes
work to the various channels by keeping storage space within each
channel and an extra space that is given to the client. Each time the
client fills the space with data and calls into multifd, that space is
given to the next idle channel and a free storage space is taken from
the channel and given to client for the next iteration.

This means we always need (#multifd_channels + 1) memory slots to
operate multifd.

This is fine, except that the presence of this one extra memory slot
doesn't allow different types of payloads to be processed at the same
time in different channels, i.e. the data type of
multifd_send_state->pages needs to be the same as p->pages.

For each new data type different from MultiFDPage_t that is to be
handled, this logic would need to be duplicated by adding new fields
to multifd_send_state, to the channels and to multifd_send_pages().

Fix this situation by moving the extra slot into the client and using
only the generic type MultiFDSendData in the multifd core.

Reviewed-by: Peter Xu <peterx@redhat.com>
Signed-off-by: Fabiano Rosas <farosas@suse.de>
(cherry picked from commit a71ef5c7f329533a49ab164b92945267be864ede)

JIRA: https://jira.oci.oraclecorp.com/browse/LINUX-15752

Conflicts:
	migration/multifd.c: No commit 9346fa187078
	("migration/multifd: Remove QEMUFile from where it is not
	needed"), contextual due to missing commits 51b07548f7c3
	("migration: Move migrate_use_multifd() to options.c"),
	11dd7be57524 ("migration/multifd: Remove p->quit from recv side"),
	06833d83f897 ("migration/multifd: Allow multifd without packets").

	migration/ram.c: No commit c323518a7aab ("migration:
	Initialize and cleanup decompression in migration.c").

Reviewed-by: Liam Merwick <liam.merwick@oracle.com>
Signed-off-by: Maciej S. Szmigiero <maciej.szmigiero@oracle.com>