block: fix failing assert on paused VM migration
Instead of throwing an assert let's just ignore that flag is already set
and return.  We assume that it's going to be safe to ignore.  Otherwise
this assert fails when migrating a paused VM back and forth.

Ideally we'd like to have a more sophisticated solution, e.g. not even
scan the nodes which should be inactive at this point.

Signed-off-by: Andrey Drobyshev <andrey.drobyshev@virtuozzo.com>

Orabug: 37106834

Link:
https://lore.kernel.org/qemu-devel/20240924125611.664315-2-andrey.drobyshev@virtuozzo.com/
Note: This is a workaround to prevent Qemu crashes and it should not be
forwarded ported until full solution is figured out.

Signed-off-by: Elena Ufimtseva <elena.ufimtseva@oracle.com>
Reviewed-by: Joao Martins <joao.m.martins@oracle.com>