migration: Emit event when source starts switchover
Add an event that is emitted while the guest is running (i.e. right
before we stop devices and vcpus) for destination software to have
more time to prepare the destination VM. 'complete' and 'pass' aren't
enough as it unbounded or too late to establish the setup.

Reuse the SRC_DOWNTIME command but with special value of downtime = 0
to announce it's running and is about to switchover. Any other value
means the guest is not running.

Orabug: 37097503

Signed-off-by: Joao Martins <joao.m.martins@oracle.com>
Tested-by: Yicheng Wang <yicheng.wang@oracle.com>
Reviewed-by: Karl Heubaum <karl.heubaum@oracle.com>